% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_outliers.R
\name{map_outliers}
\alias{map_outliers}
\title{View Whittaker outliers on an interactive map}
\usage{
map_outliers(tp, xy, validate = TRUE, basemaps = c("Esri.WorldImagery",
  "OpenTopoMap", "Esri.WorldShadedRelief", "OpenStreetMap.Mapnik",
  "CartoDB.Positron"), ...)
}
\arguments{
\item{tp}{A two column \code{matrix}, \code{data.frame} or \code{data.table} object.
First column must refer to mean annual temperatures in Celsius degrees.
Second column must refer to annual precipitations in cm.
This two-column table represents all temperature-precipitation pairs (rows)
to be tested if they fall inside or outside of the Whittaker biome polygons.}

\item{xy}{A two column \code{matrix}, \code{data.frame} or \code{data.table} object containing
the spatial coordinates corresponding to the temperature-precipitation
pairs (rows) from the \code{tp} table above.
First column must be WGS84 longitude/easting.
Second column must be WGS84 latitude/northing.}

\item{validate}{Should the input be validated?
Variable containing a single logical constant (\code{TRUE} or \code{FALSE}).
Default is \code{TRUE}, meaning that some input checking is carried out by default.
Passed to \code{get_outliers}. See \code{help(get_outliers)} for details.
Checks also if \code{xy} is a two column \code{matrix}, \code{data.frame} or \code{data.table},
if its columns are of \code{numeric} type and if \code{xy} and \code{tp} tables
have the same number of rows.}

\item{basemaps}{Passed to \code{mapview::mapviewOptions}.
Default values are:
basemaps = c("Esri.WorldImagery",
"OpenTopoMap",
"Esri.WorldShadedRelief",
"OpenStreetMap.Mapnik",
"CartoDB.Positron")}

\item{...}{Other arguments passed to \code{mapview::mapView()}.}
}
\value{
Interactive view (map) of Whittaker outliers
as produced by \code{mapview::mapView()} function.
}
\description{
This function makes use of the \code{mapview} package functionality to produce
interactive views of temperature-precipitation Whittaker outliers.
}
\author{
Valentin Stefan
}
