% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ricklefs_colors}
\alias{Ricklefs_colors}
\title{Biome colors from \emph{Ricklefs (2008)}}
\format{Named character vector}
\usage{
Ricklefs_colors
}
\description{
Colors used for biome polygons in Figure 5.5 from \emph{Ricklefs, R. E. (2008),
The economy of nature. W. H. Freeman and Company.}
(Chapter 5, Biological Communities, The biome concept).
}
\examples{
library(plotbiomes)
library(ggplot2)
ggplot() +
 geom_polygon(data = Whittaker_biomes,
              aes(x      = temp_c,
                  y      = precp_cm,
                  fill   = biome),
              colour = "gray98", # colour of polygon border
              size   = 0.5) +    # thickness of polygon border
 # fill the polygons with predefined colors
 scale_fill_manual(name   = "Whittaker biomes",
                   breaks = names(Ricklefs_colors),
                   labels = names(Ricklefs_colors),
                   values = Ricklefs_colors)

# Run example in console with: example(Ricklefs_colors)
}
\keyword{datasets}
