% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_figure.R
\name{whittaker_base_plot}
\alias{whittaker_base_plot}
\title{Create the Whittaker biome base figure}
\usage{
whittaker_base_plot(color_palette = NULL, extra_features = NULL)
}
\arguments{
\item{color_palette}{A named vector of length 9 that contains either
color names or values. The names should correspond to biome names in the
\code{Whittaker_biomes} data object. See details for additional information.
The default is to use the colors from Figure 5.5 in Ricklefs, R. E. (2008),
\emph{The economy of nature} (Chapter 5, Biological Communities, The biome concept).}

\item{extra_features}{Additional arguments to customize the Whittaker biome plot}
}
\value{
An object of class \code{gg} and \code{ggplot}.
}
\description{
Creates the Whittaker biome figure from the vignette example.
This can be modified by passing additional \code{\link{ggplot2}} style
arguments to it
}
\details{
To specify your own color palette, create a named vector
where the names correspond to biome type and the values correspond to the
colors you'd like to use. This can either be numeric (e.g. 1,2,3, etc) or
character (e.g. 'red' or '#C1E1DD'). The names for each biome are as follows:
\itemize{
\item{\code{Tundra}}
\item{\code{Boreal forest}}
\item{\code{Temperate seasonal forest}}
\item{\code{Temperate rain forest}}
\item{\code{Tropical rain forest}}
\item{\code{Tropical seasonal forest/savanna}}
\item{\code{Subtropical desert}}
\item{\code{Temperate grassland/desert}}
\item{\code{Temperate grassland/desert}}
\item{\code{Woodland/shrubland}}
}
}
\author{
Valentin Stefan, Sam Levin
}
