% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Whittaker_biomes}
\alias{Whittaker_biomes}
\title{Whittaker's biomes - data.frame}
\format{A data frame with 775 rows and 4 variables:
\tabular{rlllllr}{
\tab \strong{Variable} \tab   \tab \strong{Type}    \tab \tab \strong{Description} \cr
[, 1] \tab \strong{temp_c}   \tab , \tab \emph{numeric}   \tab : \tab Mean annual temperature (degree Celsius)\cr
[, 2] \tab \strong{precp_cm} \tab , \tab \emph{numeric}   \tab : \tab Mean annual precipitation (cm)\cr
[, 3] \tab \strong{biome_id} \tab , \tab \emph{numeric}   \tab : \tab Biome's id\cr
[, 4] \tab \strong{biome}    \tab , \tab \emph{character} \tab : \tab Biome's name
}}
\usage{
Whittaker_biomes
}
\description{
A data.frame containing boundaries of biomes as mean annual
temperature and precipitation. Digitized from \strong{Figure 5.5} in
\emph{Ricklefs, R. E. (2008). The economy of nature. W. H. Freeman and Company.}
(Chapter 5, Biological Communities, The biome concept).
}
\details{
Values in \strong{temp_c} and \strong{precp_cm} represent the vertices of the borders
between biome polygons as they were digitized.
For more details see \href{https://rawgit.com/valentinitnelav/plotbiomes/master/html/Whittaker_biomes_dataset.html}{Whittaker_biomes_dataset}
}
\examples{
library(plotbiomes)
library(ggplot2)
ggplot() +
 geom_polygon(data = Whittaker_biomes,
              aes(x      = temp_c,
                  y      = precp_cm,
                  fill   = biome),
              colour = "gray98", # colour of polygon border
              size   = 0.5)      # thickness of polygon border

# Run example in console with: example(Whittaker_biomes)
}
\keyword{datasets}
