% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Whittaker_biomes_poly}
\alias{Whittaker_biomes_poly}
\title{Whittaker's biomes - SpatialPolygonsDataFrame}
\format{An object of class SpatialPolygonsDataFrame with 9 rows and 2 columns:
\tabular{rlllllr}{
\tab \strong{Variable} \tab   \tab \strong{Type}    \tab \tab \strong{Description} \cr
[, 1] \tab \strong{biome_id} \tab , \tab \emph{numeric}   \tab : \tab Biome's id\cr
[, 2] \tab \strong{biome}    \tab , \tab \emph{character} \tab : \tab Biome's name
}}
\usage{
Whittaker_biomes_poly
}
\description{
Spatial polygons with coordinates as mean annual temperature and precipitation,
without a CRS, representing the boundaries of biomes
digitized from \strong{Figure 5.5} in
\emph{Ricklefs, R. E. (2008). The economy of nature. W. H. Freeman and Company.}
(Chapter 5, Biological Communities, The biome concept).
}
\seealso{
\code{\link{Whittaker_biomes}} for the data.frame version
(which is required by \code{\link[ggplot2]{ggplot}}).
}
\keyword{datasets}
