% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbiomes.R
\docType{package}
\name{plotbiomes}
\alias{plotbiomes}
\alias{plotbiomes-package}
\title{Plot Whittaker biomes with ggplot2}
\description{
A package for plotting Whittaker biomes with ggplot2.
It also allows for outlier identification and mapping.

The original graph is Figure 5.5 in
\emph{Ricklefs, R. E. (2008), The economy of nature. W. H. Freeman and Company.}
(Chapter 5, Biological Communities, The biome concept).
The figure was processed and brought into an R friendly format.
Details are given in \href{https://rawgit.com/valentinitnelav/plotbiomes/master/html/Whittaker_biomes_dataset.html}{Whittaker_biomes_dataset}.
}
\section{Help}{

Examples at: \href{https://rawgit.com/valentinitnelav/plotbiomes/master/html/Whittaker_biomes_examples.html}{Whittaker_biomes_examples}

GitHub page \href{https://github.com/valentinitnelav/plotbiomes}{here}
}

\author{
Valentin Stefan, Sam Levin
}
