% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_figure.R
\name{whittaker_base_plot}
\alias{whittaker_base_plot}
\title{Create the Whittaker biome base figure}
\usage{
whittaker_base_plot(extra_features = NULL)
}
\arguments{
\item{extra_features}{Additional arguments to customize the Whittaker biome plot}
}
\value{
An object of class \code{gg} and \code{ggplot}.
}
\description{
Creates the Whittaker biome figure from the vignette example.
This can be modified by passing additional \code{\link{ggplot2}} style
arguments to it
}
\author{
Sam Levin, Valentin Stefan
}
